refactor(docs): listbox dx
